updating test storage

+++ tests/test_models/test_engine/test_file_storage.py
@@ -87,7 +87,8 @@ class TestFileStorage(unittest.TestCase):
 
     def test_reload_nonexistent_file(self):
         """Test reload method with a nonexistent file"""
-        os.remove(FileStorage._FileStorage__file_path)
+        if FileStorage._FileStorage__file_path:
+            os.remove(FileStorage._FileStorage__file_path)
         new_storage = FileStorage()
         new_storage.reload()
         all_objs = new_storage.all()
